+++ src/main.js
@@ -2,12 +2,14 @@ require("prototype.creep");
 require("prototype.spawn");
 require("prototype.tower");
 
+const REPORT = true
+
 const reporter = require("reporter");
 
 module.exports.loop = function() {
   "use strict";
 
-  if (Game.time % 100 == 0) {
+  if (REPORT && Game.time % 100 == 0) {
     const rooms = {}
     for (let key in Game.rooms) {
       rooms[key] = Game.rooms[key];